chore: 스타일과 컴포넌트 코드를 분리하여 디렉토리 분리
