+++ src/components/Cart/Receipt/Receipt.styled.ts
@@ -1,0 +1,27 @@
+import { css } from '@emotion/react';
+
+export const priceWrapper = css`
+  display: flex;
+  justify-content: space-between;
+  height: 42px;
+`;
+
+export const priceDescription = css`
+  display: flex;
+  align-items: center;
+
+  height: 100%;
+  font-size: 16px;
+  font-weight: 700;
+  color: #0a0d13;
+`;
+
+export const priceText = css`
+  display: flex;
+  align-items: center;
+
+  height: 100%;
+  font-size: 24px;
+  font-weight: 700;
+  color: #000;
+`;
