+++ src/components/common/GuideText/GuideText.styled.ts
@@ -1,0 +1,23 @@
+import { css } from '@emotion/react';
+
+export const freeDeliveryGuideWrapper = css`
+  display: flex;
+  align-items: center;
+  gap: 4px;
+
+  height: 16px;
+`;
+
+export const infoIcon = css`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  width: 16px;
+  height: 16px;
+`;
+
+export const freeDeliveryGuide = css`
+  font-size: 12px;
+  font-weight: 500;
+  color: #0a0d13;
+`;
