+++ src/components/common/LoadingFallback/LoadingFallback.styled.ts
@@ -1,0 +1,8 @@
+import { css } from '@emotion/react';
+
+export const loadingContainer = css`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex: 1;
+`;
