+++ src/components/common/LoadingFallback/LoadingFallback.tsx
@@ -1,0 +1,7 @@
+import { loadingContainer } from './LoadingFallback.styled';
+
+const LoadingFallback = () => {
+  return <div css={loadingContainer}>Loading...</div>;
+};
+
+export default LoadingFallback;
