+++ src/pages/CartPage.tsx
@@ -4,9 +4,9 @@ import { ErrorBoundary } from 'react-error-boundary';
 import { useNavigate } from 'react-router-dom';
 
 import RandomAddButton from '@/components/common/Button/RandomAddButton';
-import ErrorComponent from '@/components/common/ErrorComponent';
-import Header from '@/components/common/Header';
-import LoadingComponent from '@/components/common/LoadingComponent';
+import ErrorComponent from '@/components/common/ErrorFallback/ErrorFallback';
+import Header from '@/components/common/Header/Header';
+import LoadingComponent from '@/components/common/LoadingFallback/LoadingFallback';
 import { THEME } from '@/constants/theme';
 import CartMainSection from '@components/Cart/CartMainSection';
 import OrderConfirmButton from '@components/Cart/OrderConfirmButton';
