+++ src/pages/OrderConfirmPage.tsx
@@ -4,11 +4,11 @@ import { ErrorBoundary } from 'react-error-boundary';
 import { useNavigate } from 'react-router-dom';
 
 import PurchaseButton from '@/components/common/Button/PurchaseButton';
-import ErrorComponent from '@/components/common/ErrorComponent';
+import ErrorComponent from '@/components/common/ErrorFallback/ErrorFallback';
+import Header from '@/components/common/Header/Header';
+import LoadingComponent from '@/components/common/LoadingFallback/LoadingFallback';
 import CartConfirmMainSection from '@/components/Order/OrderConfirmMainSection';
 import { BACK_ARROW } from '@assets/images';
-import Header from '@components/common/Header';
-import LoadingComponent from '@components/common/LoadingComponent';
 
 export default function OrderConfirmPage() {
   const navigate = useNavigate();
