+++ src/pages/PurchaseConfirmPage.tsx
@@ -1,10 +1,10 @@
 import { Suspense } from 'react';
 import { ErrorBoundary } from 'react-error-boundary';
 
+import ErrorComponent from '@/components/common/ErrorFallback/ErrorFallback';
+import Header from '@/components/common/Header/Header';
+import LoadingComponent from '@/components/common/LoadingFallback/LoadingFallback';
 import ReturnCartButton from '@components/common/Button/ReturnCartButton';
-import ErrorComponent from '@components/common/ErrorComponent';
-import Header from '@components/common/Header';
-import LoadingComponent from '@components/common/LoadingComponent';
 import PurchaseConfirmMainSection from '@components/Purchase/PurchaseConfirmMainSection';
 
 export default function PurchaseConfirmPage() {
